server/requirements.txt
flask
watchdog
openpyxl
ai21
python-dotenv
pymongo
fpdf
zerobouncesdk
google-genai
pandas
langchain
langchain-community
langchain_chroma
python-Levenshtein
langchain_google_vertexai
pypdf
Flask-Cors